--- Kontent.Ai.Delivery.Abstractions/ContentItems/Universal/IUniversalModelProvider.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Kontent.Ai.Delivery.Abstractions
-{
-    public interface IUniversalItemModelProvider
-    {
-        public Task<IUniversalContentItem> GetContentItemGenericModelAsync(object item);
-    }
-}
