+++ Kontent.Ai.Delivery.Caching/DeliveryClientCache.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Kontent.Ai.Delivery.Abstractions;
+using Microsoft.Extensions.Options;
 
 namespace Kontent.Ai.Delivery.Caching
 {
@@ -13,6 +14,10 @@ namespace Kontent.Ai.Delivery.Caching
     {
         private readonly IDeliveryClient _deliveryClient;
         private readonly IDeliveryCacheManager _deliveryCacheManager;
+
+        public IOptionsMonitor<DeliveryOptions> DeliveryOptions => _deliveryClient.DeliveryOptions;
+
+        public object Serializer => _deliveryClient.Serializer;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DeliveryClientCache"/> class for retrieving cached content of the specified project.
@@ -157,16 +162,9 @@ namespace Kontent.Ai.Delivery.Caching
                 CacheHelpers.GetLanguagesDependencies);
         }
 
-        // TODO
-        public Task<IDeliveryUniversalItemResponse> GetUniversalItemAsync(string codename, IEnumerable<IQueryParameter> parameters = null)
+        public Task<IApiResponse> GetDeliveryResponseAsync(string endpointUrl, string continuationToken = null)
         {
-            throw new NotImplementedException();
-        }
-
-        // TODO
-        public Task<IDeliveryUniversalItemListingResponse> GetUniversalItemsAsync(IEnumerable<IQueryParameter> parameters = null)
-        {
-            throw new NotImplementedException();
+            return _deliveryClient.GetDeliveryResponseAsync(endpointUrl, continuationToken);
         }
     }
 }
