+++ Kontent.Ai.Delivery.Extensions.Universal/DeliveryItemGenericResponse.cs
@@ -2,15 +2,17 @@ using System.Collections.Generic;
 using Kontent.Ai.Delivery.Abstractions;
 using Kontent.Ai.Delivery.SharedModels;
 
-namespace Kontent.Ai.Delivery.ContentItems.Universal
+namespace Kontent.Ai.Delivery.Extensions.Universal
 {
-    internal class DeliveryUniversalItemResponse : AbstractResponse, IDeliveryUniversalItemResponse
+    internal class DeliveryUniversalItemResponse : IResponse, IDeliveryUniversalItemResponse
     {
         public IUniversalContentItem Item { get; }
 
         public Dictionary<string, IUniversalContentItem> LinkedItems { get; }
 
-        public DeliveryUniversalItemResponse(IApiResponse response) : base(response)
+        public IApiResponse ApiResponse { get; }
+
+        public DeliveryUniversalItemResponse(IApiResponse response)
         {
             ApiResponse = response;
         }
