+++ Kontent.Ai.Delivery/ContentItems/DeliveryItemListingResponse.cs
@@ -27,7 +27,7 @@ namespace Kontent.Ai.Delivery.ContentItems
         /// <param name="items">A collection of content items of a specific type.</param>
         /// <param name="pagination">Response paging information.</param>
         [JsonConstructor]
-        internal DeliveryItemListingResponse(ApiResponse response, IList<T> items,IPagination pagination) : base(response)
+        internal DeliveryItemListingResponse(IApiResponse response, IList<T> items,IPagination pagination) : base(response)
         {
             Items = items;
             Pagination = pagination;
@@ -37,7 +37,7 @@ namespace Kontent.Ai.Delivery.ContentItems
         /// Initializes a new instance of the <see cref="DeliveryItemListingResponse{T}"/> class.
         /// </summary>
         /// <param name="response">The response from Kontent.ai Delivery API that contains content items.</param>
-        internal DeliveryItemListingResponse(ApiResponse response) : base(response)
+        internal DeliveryItemListingResponse(IApiResponse response) : base(response)
         {
         }
     }
