+++ Kontent.Ai.Delivery/ContentItems/Elements/AssetElementValue.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class AssetElementValue : ContentElementValue<IEnumerable<IAsset>>, IAssetElementValue
+    public class AssetElementValue : ContentElementValue<IEnumerable<IAsset>>, IAssetElementValue
     {
     }
 }
