+++ Kontent.Ai.Delivery/ContentItems/Elements/DateTimeElementValue.cs
@@ -4,7 +4,7 @@ using System;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class DateTimeElementValue : ContentElementValue<DateTime?>, IDateTimeElementValue
+    public class DateTimeElementValue : ContentElementValue<DateTime?>, IDateTimeElementValue
     {
         [JsonProperty("display_timezone")]
         public string DisplayTimezone { get; set; }
