+++ Kontent.Ai.Delivery/ContentItems/Elements/LinkedItemsElementValue.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class LinkedItemsElementValue: ContentElementValue<IEnumerable<string>>
+    public class LinkedItemsElementValue: ContentElementValue<IEnumerable<string>>
     {
         
     }
