+++ Kontent.Ai.Delivery/ContentItems/Elements/MultipleChoiceElementValue.cs
@@ -3,7 +3,7 @@ using Kontent.Ai.Delivery.Abstractions;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class MultipleChoiceElementValue:  ContentElementValue<IEnumerable<IMultipleChoiceOption>>
+    public class MultipleChoiceElementValue:  ContentElementValue<IEnumerable<IMultipleChoiceOption>>
     {
         
     }
