+++ Kontent.Ai.Delivery/ContentItems/Elements/NumberElementValue.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class NumberElementValue: ContentElementValue<decimal?>
+    public class NumberElementValue: ContentElementValue<decimal?>
     {
     }
 }
