+++ Kontent.Ai.Delivery/ContentItems/Elements/StringElementValue.cs
@@ -1,6 +1,6 @@
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class StringElementValue : ContentElementValue<string>
+    public class StringElementValue : ContentElementValue<string>
     {
     }
 }
