+++ Kontent.Ai.Delivery/ContentItems/Elements/TaxonomyElementValue.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class TaxonomyElementValue : ContentElementValue<IEnumerable<ITaxonomyTerm>>, ITaxonomyElementValue
+    public class TaxonomyElementValue : ContentElementValue<IEnumerable<ITaxonomyTerm>>, ITaxonomyElementValue
     {
         [JsonProperty("taxonomy_group")]
         public string TaxonomyGroup { get; set; }
