+++ Kontent.Ai.Delivery/ContentItems/Elements/TextElementValue.cs
@@ -1,6 +1,6 @@
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class TextElementValue : StringElementValue
+    public class TextElementValue : StringElementValue
     {
 
     }
