+++ Kontent.Ai.Delivery/ContentItems/Elements/UrlSlugElementValue.cs
@@ -1,6 +1,6 @@
 namespace Kontent.Ai.Delivery.ContentItems.Elements
 {
-    internal class UrlSlugElementValue : StringElementValue
+    public class UrlSlugElementValue : StringElementValue
     {
 
     }
