+++ Kontent.Ai.Delivery/ContentTypes/DeliveryTypeListingResponse.cs
@@ -27,7 +27,7 @@ namespace Kontent.Ai.Delivery.ContentTypes
         /// <param name="types">A collection of content types.</param>
         /// <param name="pagination">Response paging information.</param>
         [JsonConstructor]
-        internal DeliveryTypeListingResponse(ApiResponse response, IList<IContentType> types, IPagination pagination) : base(response)
+        internal DeliveryTypeListingResponse(IApiResponse response, IList<IContentType> types, IPagination pagination) : base(response)
         {
             Types = types;
             Pagination = pagination;
@@ -37,7 +37,7 @@ namespace Kontent.Ai.Delivery.ContentTypes
         /// Initializes a new instance of the <see cref="DeliveryTypeListingResponse"/> class.
         /// </summary>
         /// <param name="response">The response from Kontent.ai Delivery API that contains content types.</param>
-        internal DeliveryTypeListingResponse(ApiResponse response) : base(response)
+        internal DeliveryTypeListingResponse(IApiResponse response) : base(response)
         {
         }
     }
