+++ Kontent.Ai.Delivery/Languages/DeliveryLanguageListingResponse.cs
@@ -27,7 +27,7 @@ namespace Kontent.Ai.Delivery.Languages
         /// <param name="languages">A collection of languages.</param>
         /// <param name="pagination">Response paging information.</param>
         [JsonConstructor]
-        internal DeliveryLanguageListingResponse(ApiResponse response, IList<ILanguage> languages, IPagination pagination) : base(response)
+        internal DeliveryLanguageListingResponse(IApiResponse response, IList<ILanguage> languages, IPagination pagination) : base(response)
         {
             Languages = languages;
             Pagination = pagination;
@@ -37,7 +37,7 @@ namespace Kontent.Ai.Delivery.Languages
         /// Initializes a new instance of the <see cref="DeliveryLanguageListingResponse"/> class.
         /// </summary>
         /// <param name="response">The response from Kontent.ai Delivery API that contains languages.</param>
-        internal DeliveryLanguageListingResponse(ApiResponse response) : base(response)
+        internal DeliveryLanguageListingResponse(IApiResponse response) : base(response)
         {
         }
     }
