+++ Kontent.Ai.Delivery/SharedModels/ApiResponse.cs
@@ -13,7 +13,7 @@ namespace Kontent.Ai.Delivery.SharedModels
     /// Represents a successful JSON response from Kontent.ai Delivery API.
     /// </summary>
     [DebuggerDisplay("Url = {" + nameof(RequestUrl) + "}")]
-    internal sealed class ApiResponse : IApiResponse
+    public sealed class ApiResponse : IApiResponse
     {
         private JObject _jsonContent;
         private string _content;
@@ -48,7 +48,7 @@ namespace Kontent.Ai.Delivery.SharedModels
 
         /// <inheritdoc/>
         public IError Error { get; }
-        
+
         /// <summary>
         /// Initializes a new instance of the <see cref="ApiResponse"/> class.
         /// </summary>
@@ -56,7 +56,7 @@ namespace Kontent.Ai.Delivery.SharedModels
         /// <param name="hasStaleContent">Specifies whether content is stale.</param>
         /// <param name="continuationToken">Continuation token to be used for continuing enumeration.</param>
         /// <param name="requestUrl">URL used to retrieve this response.</param>
-        internal ApiResponse(HttpContent httpContent, bool hasStaleContent, string continuationToken, string requestUrl) : 
+        internal ApiResponse(HttpContent httpContent, bool hasStaleContent, string continuationToken, string requestUrl) :
             this(httpContent, hasStaleContent, continuationToken, requestUrl, null)
         {
         }
@@ -100,6 +100,11 @@ namespace Kontent.Ai.Delivery.SharedModels
                 _jsonContent = await JObject.LoadAsync(jsonReader);
             }
             return _jsonContent;
+        }
+
+        async Task<object> IApiResponse.GetJsonContentAsync()
+        {
+            return await Task.FromResult((object)GetJsonContentAsync());
         }
     }
 }
