+++ Kontent.Ai.Delivery/TaxonomyGroups/DeliveryTaxonomyListingResponse.cs
@@ -33,7 +33,7 @@ namespace Kontent.Ai.Delivery.TaxonomyGroups
         /// <param name="taxonomies">A collection of taxonomies.</param>
         /// <param name="pagination">Response paging information.</param>
         [JsonConstructor]
-        internal DeliveryTaxonomyListingResponse(ApiResponse response, IList<ITaxonomyGroup> taxonomies, IPagination pagination) : base(response)
+        internal DeliveryTaxonomyListingResponse(IApiResponse response, IList<ITaxonomyGroup> taxonomies, IPagination pagination) : base(response)
         {
             Taxonomies = taxonomies;
             Pagination = pagination;
@@ -43,7 +43,7 @@ namespace Kontent.Ai.Delivery.TaxonomyGroups
         /// Initializes a new instance of the <see cref="DeliveryTaxonomyListingResponse"/> class.
         /// </summary>
         /// <param name="response">The response from Kontent.ai Delivery API that contains taxonomies.</param>
-        internal DeliveryTaxonomyListingResponse(ApiResponse response) : base(response)
+        internal DeliveryTaxonomyListingResponse(IApiResponse response) : base(response)
         {
         }
     }
